refactor: 디렉토리 구조 정리
- 디렉토리 구조 정리
- 게시판 API 테스트 완료
- swagger parameter 추가
