+++ src/main/java/com/example/homeaid/admin/adminreply/entity/PostType.java
@@ -1,0 +1,10 @@
+package com.example.homeaid.admin.adminreply.entity;
+
+import lombok.Getter;
+
+@Getter
+public enum PostType {
+  CUSTOMER,
+  MANAGER
+
+}
