--- src/main/java/com/example/homeaid/adminreply/entity/PostType.java
@@ -1,10 +1,0 @@
-package com.example.homeaid.adminreply.entity;
-
-import lombok.Getter;
-
-@Getter
-public enum PostType {
-  CUSTOMER,
-  MANAGER
-
-}
