--- src/main/java/com/example/homeaid/customer/entity/CustomerAddress.java
@@ -1,41 +1,0 @@
-package com.example.homeaid.customer.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-@Table(name = "customer_address")
-public class CustomerAddress {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @Setter
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "customer_id")
-  private Customer customer;
-
-  private String address;
-
-  private String addressDetail;
-
-  @Builder
-  public CustomerAddress(String address, String addressDetail) {
-    this.address = address;
-    this.addressDetail = addressDetail;
-  }
-}
