--- src/main/java/com/example/homeaid/managerboard/dto/response/UpdateManagerBoardResponseDto.java
@@ -1,27 +1,0 @@
-package com.example.homeaid.managerboard.dto.response;
-
-import com.example.homeaid.managerboard.entity.ManagerBoard;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class UpdateManagerBoardResponseDto {
-
-  private Long id;
-  private String title;
-  private String content;
-
-  public static UpdateManagerBoardResponseDto toDto(ManagerBoard managerBoard) {
-    return UpdateManagerBoardResponseDto.builder()
-        .id(managerBoard.getId())
-        .title(managerBoard.getTitle())
-        .content(managerBoard.getContent())
-        .build();
-  }
-
-}
