--- src/main/java/com/example/homeaid/managerboard/entity/ManagerBoard.java
@@ -1,48 +1,0 @@
-package com.example.homeaid.managerboard.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import java.time.LocalDateTime;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@Entity
-@Getter
-@Builder
-@Table(name = "manager_board")
-@NoArgsConstructor
-@AllArgsConstructor
-@EntityListeners(AuditingEntityListener.class)
-public class ManagerBoard {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @Column(nullable = false, length = 255)
-  private String title;
-  @Column(nullable = false, columnDefinition = "TEXT")
-  private String content;
-
-  private Boolean isPublic; // 공개여부
-
-  @CreatedDate
-  @Column(updatable = false)
-  private LocalDateTime createdAt;
-
-//  @LastModifiedDate
-//  @Column(nullable = false)
-//  private LocalDateTime updatedAt;
-
-  private Long managerId;
-}
